+++ .yarnrc.yml
@@ -1,12 +1,2 @@
 yarnPath: .yarn/releases/yarn-3.6.1.cjs
 nodeLinker: pnp
-
-# packageExtensions:
-#   http-proxy-middleware@*:
-#     dependencies:
-#       http-proxy-middleware: "5.3.0"
-
-# proxies:
-#   "/ws":
-#     target: "http://localhost:3000"
-#     ws: true
